.github/workflows/workflow_scripts/requirements.txt
snowflake-connector-python==3.15.0
boto3==1.35.11